@@ -1,25 +1,25 @@
 module github.com/nsip/mrac-2023
 
-go 1.20
+go 1.21
 
 require (
-	github.com/digisan/go-generics v0.3.5
-	github.com/digisan/gotk v0.3.8
-	github.com/digisan/json-tool v0.2.5
-	github.com/digisan/logkit v0.2.5
-	github.com/tidwall/gjson v1.14.4
+	github.com/digisan/go-generics v0.4.0
+	github.com/digisan/gotk v0.4.1
+	github.com/digisan/json-tool v0.2.7
+	github.com/digisan/logkit v0.2.7
+	github.com/tidwall/gjson v1.16.0
 	github.com/tidwall/sjson v1.2.5
 )
 
 require (
 	github.com/BurntSushi/toml v1.3.2 // indirect
 	github.com/clbanning/mxj v1.8.4 // indirect
-	github.com/gookit/color v1.5.3 // indirect
+	github.com/gookit/color v1.5.4 // indirect
 	github.com/h2non/filetype v1.1.3 // indirect
 	github.com/tidwall/match v1.1.1 // indirect
 	github.com/tidwall/pretty v1.2.1 // indirect
 	github.com/xo/terminfo v0.0.0-20220910002029-abceb7e1c41e // indirect
-	golang.org/x/net v0.11.0 // indirect
-	golang.org/x/sys v0.9.0 // indirect
+	golang.org/x/net v0.14.0 // indirect
+	golang.org/x/sys v0.11.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
